Tests/LibWeb/Text/expected/XHR/FormData-constructor-takes-name-and-value-of-submitter.txt

'test-button-name' = 'test-button-value'
